+++ controller/controller_user.js
@@ -2,6 +2,7 @@ const mapping = require('../database/mapping.js');
 const repo = require('../database/repository.js');
 const appConst = require('../common/applicationConstant.js')
 const mongoose = require('mongoose');
+const staticFunc = require('../common/staticFunc.js');
 
 module.exports ={
 
@@ -28,6 +29,11 @@ module.exports ={
     update : async (obj)=>{
         
         try{
+
+            if (obj.filter==undefined || staticFunc.isJsonEmpty(obj.filter)) throw 'filter should not be empty';
+
+            else if (obj.value==undefined || staticFunc.isJsonEmpty(obj.value)) throw 'value should not be empty';
+
             return await repo.update(await mapping.blankUser(),obj.filter,obj.value);
         }
         catch(e){
@@ -35,8 +41,11 @@ module.exports ={
         }
 
     },
-    delete : async (condition)=>{
+    remove : async (condition)=>{
         try{
+
+            if (condition==undefined ||staticFunc.isJsonEmpty(condition)) throw 'condition should not be empty';
+
             return await repo.delete(await mapping.blankUser(),condition);
         }
         catch(e){
